requirements.txt
# Core Python Dependencies
apache-airflow-providers-google==10.15.0